+++ lib/galaxy_test/api/_framework.py
@@ -1,5 +1,19 @@
-# TODO: implement a remote-only version of ApiTestCase if galaxy_test.driver isn't
-# available. Most of these tests do not depend on Galaxy internals.
-from galaxy_test.driver.api import ApiTestCase
+from galaxy_test.base.api import UsesApiTestCaseMixin
+from galaxy_test.base.testcase import FunctionalTestCase
+try:
+    from galaxy_test.driver.driver_util import GalaxyTestDriver
+except ImportError:
+    # Galaxy libraries and galaxy test driver not available, just assume we're
+    # targetting a remote Galaxy.
+    GalaxyTestDriver = None
+
+
+class ApiTestCase(FunctionalTestCase, UsesApiTestCaseMixin):
+    galaxy_driver_class = GalaxyTestDriver
+
+    def setUp(self):
+        super(ApiTestCase, self).setUp()
+        self._setup_interactor()
+
 
 __all__ = ('ApiTestCase', )
