+++ lib/galaxy_test/base/env.py
@@ -1,7 +1,9 @@
 """Base utilities for working Galaxy test environments.
 """
+import fcntl
 import os
 import socket
+import struct
 
 DEFAULT_WEB_HOST = socket.gethostbyname('localhost')
 
@@ -22,3 +24,12 @@ def target_url_parts():
     default_url = "http://%s:%s" % (host, port)
     url = os.environ.get('GALAXY_TEST_EXTERNAL', default_url)
     return host, port, url
+
+
+def get_ip_address(ifname):
+    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    return socket.inet_ntoa(fcntl.ioctl(
+        s.fileno(),
+        0x8915,  # SIOCGIFADDR
+        struct.pack('256s', ifname[:15].encode('utf-8'))
+    )[20:24])
